fix: avoid holding character locks across awaits
- Wake async_shared_mutex waiters outside the unlock mutex
- Move characters off synchronized<> onto an internal shared_mutex
- Snapshot privately under lock; fan out via foreach without holding it
- Expose collect_ids for system mail/storage id lists
- Convert listeners and related game APIs to awaitable send paths
- Update AMQP, timers, builtins, and group/clan call sites

+++ include/fb/game/protocol/character/update_external.h
@@ -144,7 +144,7 @@ public:
 
             if (ch.items.armor() != nullptr)
             {
-                serializer.appearance.armor = ch.items.armor()->based<fb::model::armor>().dress;
+                serializer.appearance.armor = static_cast<uint8_t>(ch.items.armor()->based<fb::model::armor>().dress);
                 if (serializer.appearance.armor_color.has_value() == false)
                     serializer.appearance.armor_color = ch.items.armor()->based<fb::model::armor>().color;
             }
